requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiogram_dialog==2.3.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
attrs==24.3.0
billiard==4.2.1
black==25.1.0
boto3==1.36.12
botocore==1.36.12
cachetools==5.5.0
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dateparser==1.2.0
debugpy==1.8.12
docker==7.1.0
frozenlist==1.5.0
idna==3.10
iniconfig==2.0.0
intervaltree==3.1.0
Jinja2==3.1.5
jmespath==1.0.1
kombu==5.4.2
loguru==0.7.3
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
peewee==3.17.8
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.2.1
psutil==6.1.1
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
pytest==8.3.4
pytest-celery==1.1.3
pytest-docker-tools==3.1.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-memcached==1.62
pytz==2024.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
sortedcontainers==2.4.0
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
